requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Frozen-Flask==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
Pygments==2.17.2
PyYAML==6.0.1
Werkzeug==3.0.1
